device level config
