--- k8s-deployment.yml
@@ -1,55 +1,0 @@
----
-apiVersion: v1
-kind: Service
-metadata:
-  name: sensors-honeycomb-api
-  annotations:
-    external-dns.alpha.kubernetes.io/hostname: honeycomb.apparatus.wildflowertesting.org
-    service.beta.kubernetes.io/aws-load-balancer-ssl-cert: arn:aws:acm:us-east-2:204031725010:certificate/929c762f-e3c7-45ff-a5ac-bd482682d294
-    service.beta.kubernetes.io/aws-load-balancer-backend-protocol: http
-spec:
-  type: LoadBalancer
-  ports:
-  - port: 443
-    name: http
-    targetPort: 4000
-  selector:
-    app: sensors-honeycomb-api
-
----
-apiVersion: extensions/v1beta1
-kind: Deployment
-metadata:
-  name: honeycomb
-spec:
-  template:
-    metadata:
-      labels:
-        app: sensors-honeycomb-api
-    spec:
-      containers:
-      - image: "wildflowerschools/sensors_honeycomb_api:v2019-01-15-15-16-15"
-        name: sensors-honeycomb-api
-        env:
-          - name: ENV
-            value: "test"
-          - name: PGPASSWORD
-            valueFrom:
-              configMapKeyRef:
-                name: honeycomb-postgres-config
-                key: POSTGRES_PASSWORD
-          - name: PGUSER
-            valueFrom:
-              configMapKeyRef:
-                name: honeycomb-postgres-config
-                key: POSTGRES_USER
-          - name: PGDATABASE
-            valueFrom:
-              configMapKeyRef:
-                name: honeycomb-postgres-config
-                key: POSTGRES_DB
-          - name: PGHOST
-            value: honeycomb-postgres.default.svc.cluster.local
-        ports:
-        - containerPort: 4000
-          name: http
