--- k8s-postgres.yml
@@ -1,99 +1,0 @@
----
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: honeycomb-postgres-config
-  labels:
-    app: honeycomb-postgres
-data:
-  POSTGRES_DB: honeycomb
-  POSTGRES_USER: honeycomb_app_rw
-  POSTGRES_PASSWORD: D0BF56B00DA62700941259C30607C39D
-
----
-kind: StorageClass
-apiVersion: storage.k8s.io/v1
-metadata:
-  name: postgres-storage
-provisioner: kubernetes.io/aws-ebs
-reclaimPolicy: Retain
-parameters:
-  type: gp2
-  iopsPerGB: "10"
-  fsType: ext4
-
----
-kind: PersistentVolume
-apiVersion: v1
-metadata:
-  name: postgres-pv-volume
-  labels:
-    type: ebs
-    app: honeycomb-postgres
-spec:
-  storageClassName: postgres-storage
-  capacity:
-    storage: 50Gi
-  accessModes:
-    - ReadWriteOnce
-  hostPath:
-    path: "/mnt/postgres-data"
-
-
----
-kind: PersistentVolumeClaim
-apiVersion: v1
-metadata:
-  name: postgres-pv-claim
-  labels:
-    app: honeycomb-postgres
-spec:
-  storageClassName: postgres-storage
-  accessModes:
-    - ReadWriteOnce
-  resources:
-    requests:
-      storage: 50Gi
-
----
-apiVersion: extensions/v1beta1
-kind: Deployment
-metadata:
-  name: honeycomb-postgres
-spec:
-  replicas: 1
-  template:
-    metadata:
-      labels:
-        app: honeycomb-postgres
-    spec:
-      containers:
-        - name: honeycomb-postgres
-          image: postgres:10.4
-          imagePullPolicy: "Always"
-          ports:
-            - containerPort: 5432
-          envFrom:
-            - configMapRef:
-                name: honeycomb-postgres-config
-          volumeMounts:
-            - mountPath: /var/lib/postgresql/data
-              name: honeycomb-postgres
-      volumes:
-        - name: honeycomb-postgres
-          persistentVolumeClaim:
-            claimName: postgres-pv-claim
-
----
-apiVersion: v1
-kind: Service
-metadata:
-  name: honeycomb-postgres
-  labels:
-    app: honeycomb-postgres
-spec:
-  type: NodePort
-  ports:
-   - port: 5432
-  selector:
-   app: honeycomb-postgres
