+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "honeycomb",
-    "version": "1.1.0",
+    "version": "1.1.1",
     "description": "wildflower data service",
     "repository": "https://github.com/WildflowerSchools/sensors_honeycomb_api",
     "main": "src/index.js",
@@ -18,7 +18,7 @@
     "author": "pjdecoursey",
     "license": "MIT",
     "dependencies": {
-        "@wildflowerschools/graphql-beehive": "^0.7.6",
+        "@wildflowerschools/graphql-beehive": "^0.7.7",
         "apollo": "^1.9.2",
         "apollo-server": "^2.1.0",
         "apollo-server-express": "^2.1.0",
