+++ src/schema/common.js
@@ -1,0 +1,25 @@
+exports.typeDefs = `
+  
+enum PropertyType {
+    BOOL
+    STR
+    INT
+    FLOAT
+    NULL
+}
+
+type Property {
+    name: String!
+    value: String
+    type: PropertyType!
+}
+
+
+input PropertyInput {
+    name: String!
+    value: String
+    type: PropertyType!
+}
+
+
+`;
